+++ Sources/KeyboardKitSwiftUI/Layout/KeyboardDimensions.swift
@@ -17,8 +17,16 @@ public protocol KeyboardDimensions {
     
     var buttonHeight: CGFloat { get }
     var buttonInsets: EdgeInsets { get }
+    
+    
+    @available(*, deprecated, message: "Use the new KeyboardButtonWidth-based width function instead.")
     var longButtonWidth: CGFloat { get }
+    
+    @available(*, deprecated, message: "Use the new KeyboardButtonWidth-based width function instead.")
     var shortButtonWidth: CGFloat { get }
     
+    @available(*, deprecated, message: "Use the new KeyboardButtonWidth-based width function instead.")
     func width(for action: KeyboardAction, keyboardWidth: CGFloat, context: KeyboardContext) -> CGFloat?
+    
+    func width(for action: KeyboardAction, at row: Int, rowItemCount: Int, rowItemIndex: Int, context: KeyboardContext) -> KeyboardButtonWidth
 }