+++ Sources/KeyboardKitSwiftUI/System/SystemKeyboard.swift
@@ -21,7 +21,7 @@ public struct SystemKeyboard: View {
     
     public init(
         layout: KeyboardLayout,
-        dimensions: KeyboardDimensions = SystemKeyboardDimensions(),
+        dimensions: KeyboardDimensions = StandardKeyboardDimensions(),
         buttonBuilder: @escaping ButtonBuilder = Self.standardButtonBuilder) {
         self.rows = layout.actionRows
         self.dimensions = dimensions
@@ -32,7 +32,9 @@ public struct SystemKeyboard: View {
     private let dimensions: KeyboardDimensions
     private let rows: KeyboardActionRows
     
-    @State private var size: CGSize = .zero
+    @State private var keyboardSize: CGSize = .zero
+    @State private var referenceSize: CGSize = .zero
+    
     @EnvironmentObject var context: ObservableKeyboardContext
     
     public typealias ButtonBuilder = (KeyboardAction, KeyboardSize) -> AnyView
@@ -44,7 +46,7 @@ public struct SystemKeyboard: View {
                 row(at: $0.offset, actions: $0.element)
             }
         }
-        .bindSize(to: $size)
+        .bindSize(to: $keyboardSize)
         .inputCallout(style: .systemStyle(for: context))
         .secondaryInputCallout(for: context, style: .systemStyle(for: context))
     }
@@ -63,32 +65,20 @@ public extension SystemKeyboard {
 
 private extension SystemKeyboard {
 
-    func row(at index: Int, actions: KeyboardActionRow) -> some View {
+    func row(at rowIndex: Int, actions: KeyboardActionRow) -> some View {
         HStack(spacing: 0) {
-            rowEdgeSpacer(at: index)
-            ForEach(Array(actions.enumerated()), id: \.offset) {
-                SystemKeyboardButtonRowItem(action: $0.element, buttonContent: buttonBuilder($0.element, size), keyboardSize: size)
+            ForEach(Array(actions.enumerated()), id: \.offset) { action in
+                let width = dimensions.width(for: action.element, at: rowIndex, rowItemCount: actions.count, rowItemIndex: action.offset, context: context)
+                buttonBuilder(action.element, keyboardSize)
+                    .frame(maxWidth: .infinity)
+                    .frame(height: dimensions.buttonHeight - dimensions.buttonInsets.top - dimensions.buttonInsets.bottom)
+                    .standardButtonStyle(for: action.element, context: context)
+                    .padding(dimensions.buttonInsets)
+                    .frame(height: dimensions.buttonHeight)
+                    .background(Color.clearInteractable)
+                    .width(width, totalWidth: keyboardSize.width, referenceSize: $referenceSize)
+                    .keyboardAction(action.element, actionHandler: context.actionHandler)
             }
-            rowEdgeSpacer(at: index)
         }
-    }
-    
-    @ViewBuilder
-    func rowEdgeSpacer(at index: Int) -> some View {
-        if index == 1 {
-            Spacer(minLength: secondRowPadding)
-        } else {
-            EmptyView()
-        }
-    }
-    
-    /**
-     A temp. way to get side padding on each side on English
-     iPhone keyboards.
-     */
-    var secondRowPadding: CGFloat {
-        guard Locale.current.identifier.starts(with: "en") else { return 0 }
-        guard UIDevice.current.userInterfaceIdiom == .phone else { return 0 }
-        return max(0, 20 * CGFloat(rows[0].count - rows[1].count))
     }
 }