+++ Sources/KeyboardKitSwiftUI/_Deprecations/SystemKeyboardDimensions+Deprecations.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+@available(*, deprecated, renamed: "StandardKeyboardDimensions")
+public typealias SystemKeyboardDimensions = StandardKeyboardDimensions